+++ NEWS.md
@@ -2,6 +2,8 @@ Red Panda C++ Version 2.12
 
   - fix: Can't correctly load project's custom compile options, if it contains more than one line contents.
   - fix: Crash when create or open txt files in project.
+  - enhancement: Code folding for #if/#endif
+  - enhancement: When folding "if", don't fold "else";
 
 Red Panda C++ Version 2.11